requirements.txt
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
Flask==2.3.3
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.0
referencing==0.30.2
rpds-py==0.10.3
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==2.3.7
